# src/main/resources/openapi.yaml
openapi: 3.1.0
info:
  title: OpenAPI definition
  version: v0
security:
  - sessionAuth: []
paths:
...
  /api/v1/users/me/onboarding:
    put:
      tags:
      - user
      summary: Complete user onboarding
      description: |
        Completes the onboarding flow for the currently authenticated user.
        Stores the user's selected learning path + preferences and initializes their course progress.
        Returns the updated user profile and onboarding state.
      operationId: submitOnboarding
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OnboardingSubmission"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/OnboardingResponse"
  /api/v1/users/me/avatar:
    put:
      tags:
      - user
      summary: Update user avatar
      description: |
        Updates the avatar configuration for the currently authenticated user.
        Stores the selected avatar version and index.
        Returns the updated user profile.
      operationId: changeUserAvatar
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AvatarInfo"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UserResponse"
...
  /api/v1/users:
    get:
      tags:
      - user
      summary: Get users by ID list
      description: |
        Returns user profiles for the provided list of user IDs.
      operationId: getUsersByIds
      parameters:
      - name: userIds
        in: query
        required: true
        schema:
          type: array
          items:
            type: string
            format: uuid
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserResponse"
...
components:
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: JWT
  schemas:
    OnboardingSubmission:
      type: object
      properties:
        selectedUsername:
          type: string
        chosenPath:
          type: string
          enum:
          - DATA
          - IOS
        chosenCourse:
          type: string
          format: uuid
        hasProgrammingExperience:
          type: boolean
      required:
      - selectedUsername
      - chosenCourse
      - chosenPath
      - hasProgrammingExperience
    CourseProgressResponse:
      type: object
      properties:
        userId:
          type: string
          format: uuid
        courseId:
          type: string
          format: uuid
        currentLessonId:
          type: string
          format: uuid
        moduleId:
          type: string
          format: uuid
        id:
          type: string
          format: uuid
        updatedAt:
          type: string
          format: date-time
      required:
      - courseId
      - currentLessonId
      - id
      - moduleId
      - updatedAt
      - userId
    CourseProgressResponseWithEnrolled:
      type: object
      properties:
        courseProgress:
          $ref: "#/components/schemas/CourseProgressResponse"
        enrolled:
          type: array
          items:
            type: string
            format: uuid
      required:
      - courseProgress
      - enrolled
    OnboardingResponse:
      type: object
      properties:
        refreshedUser:
          $ref: "#/components/schemas/UserResponse"
        preferences:
          $ref: "#/components/schemas/UserPreferences"
        courseProgressResponse:
          $ref: "#/components/schemas/CourseProgressResponseWithEnrolled"
      required:
      - courseProgressResponse
      - preferences
      - refreshedUser
    UserPreferences:
      type: object
      properties:
        userId:
          type: string
          format: uuid
        hasExperience:
          type: boolean
        chosenPath:
          type: string
          enum:
          - DATA
          - IOS
      required:
      - chosenPath
      - hasExperience
    UserResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        displayName:
          type: string
        avatarVersion:
          type: string
        avatarIndex:
          type: integer
          format: int32
        email:
          type: string
        createdAt:
          type: string
          format: date-time
        hasOnboarded:
          type: boolean
      required:
      - avatarIndex
      - avatarVersion
      - createdAt
      - email
      - firstName
      - hasOnboarded
      - id
      - lastName
    AvatarInfo:
      type: object
      properties:
        version:
          type: string
        index:
          type: integer
          format: int32
      required:
      - index
      - version
...